:email: setup sendgrid and a template
only problem is the mailers.scss seems not to be compiled and available
in email :( without the explicit pre-compile for development or being
included in application.

tested in dev
  Rails.application.config.assets.precompile += %w[mailers.scss] if Rails.env.development?

currently prod CSS is poluted by the mailer styles, need to fix this and
add the *= stub mailers and associated TODO:'s in the code.

This seems not a problem with a fresh rails project and it seems not to
be anything to do with the app, config, directories? even removing
sass-rails fails differently

with no sass-rails
getting
  Error performing ActionMailer::MailDeliveryJob ... ActionView::Template::Error (mailers.css):
  2.7.1/lib/ruby/gems/2.7.0/gems/sprockets-rails-3.2.2/lib/sprockets/rails/helper.rb:372:in `raise_unless_precompiled_asset'
  2.7.1/lib/ruby/gems/2.7.0/gems/sprockets-rails-3.2.2/lib/sprockets/rails/helper.rb:349:in `find_debug_asset'
  ...
  2.7.1/lib/ruby/gems/2.7.0/gems/sprockets-rails-3.2.2/lib/sprockets/rails/helper.rb:170:in `stylesheet_link_tag'
  app/views/layouts/mailer.html.erb:8

should be
  ... scssc not foud or something
  2.7.1/lib/ruby/gems/2.7.0/gems/sprockets-rails-3.2.2/lib/sprockets/rails/helper.rb:348:in `find_asset'